Added confirmation page to delete tags

+++ tag/views.py
@@ -1,7 +1,7 @@
 from django.contrib import messages
 from django.http import HttpResponseNotFound
 from django.shortcuts import render, redirect
-from tag.models import Tag, TagForm
+from tag.models import Tag, TagForm, DeleteTagForm
 from user.decorators import teacher_required
 
 
@@ -66,10 +66,30 @@ def edit(request, tag_id):
 
 @teacher_required
 def delete(request, tag_id):
-    try:
-        tag = Tag.objects.get(pk=tag_id)
-        tag.delete()
-        messages.add_message(request, messages.SUCCESS, f'Tag {tag.name} was deleted')
-        return redirect('tag:index')
-    except Tag.DoesNotExist:
-        return HttpResponseNotFound()
+    """
+    View to delete a tag from the database
+    """
+
+    if request.method == 'POST':  # Form was submitted
+        form = DeleteTagForm(request.POST)
+        if form.is_valid():
+            tag = Tag.objects.get(id=tag_id)
+            if form.cleaned_data['confirmDelete']:
+                tag.delete()
+                messages.add_message(request, messages.SUCCESS, f'The tag {tag.name} was deleted')
+            else:
+                messages.add_message(request, messages.SUCCESS, f'The tag {tag.name} was not deleted')
+            return redirect('tag:index')
+
+    else:  # GET method
+        # Render page with tag info and "Delete? Yes/No"
+        if tag_id in Tag.objects.all().values_list('id', flat=True):
+            tag = Tag.objects.all().get(id=tag_id)
+            return render(request, 'tag/delete.html',
+                          {'form': DeleteTagForm(), 'tag': tag})
+
+        else:  # In case user enters invalid URL directly
+            messages.add_message(request, messages.WARNING, 'Did not find tag')
+            return redirect('tag:index')
+
+    return redirect('tag:index')
